Refactoring of WorkflowRunReport to its own class
