+++ common-utilities/src/main/java/ca/on/oicr/pde/testing/decider/DeciderRunTestReport.java
@@ -282,50 +282,4 @@ public class DeciderRunTestReport {
 
         return t;
     }
-
-    public static class WorkflowRunReport implements Comparable<WorkflowRunReport> {
-
-        Map workflowIni = new TreeMap<>();
-        List<ReducedFileProvenanceReportRecord> files = new ArrayList<>();
-
-        public Map getWorkflowIni() {
-            return workflowIni;
-        }
-
-        public void setWorkflowIni(Map<String, String> workflowIni) {
-            this.workflowIni = new TreeMap<>(workflowIni);
-        }
-
-        public List<ReducedFileProvenanceReportRecord> getFiles() {
-            Collections.sort(files);
-            return files;
-        }
-
-        public void setFiles(List<ReducedFileProvenanceReportRecord> files) {
-            this.files.addAll(files);
-            Collections.sort(this.files);
-        }
-
-        @Override
-        public String toString() {
-            return ToStringBuilder.reflectionToString(this, ToStringStyle.SHORT_PREFIX_STYLE);
-        }
-
-        @Override
-        public int hashCode() {
-            return HashCodeBuilder.reflectionHashCode(this);
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            return EqualsBuilder.reflectionEquals(this, obj);
-        }
-
-        @Override
-        public int compareTo(WorkflowRunReport o) {
-            //TODO: optimize
-            return this.toString().compareTo(o.toString());
-        }
-    }
-
 }
